agents/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.350
langchain-community==0.0.1
langgraph==0.0.20
pydantic==2.5.0
openai==1.3.7
anthropic==0.7.8
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
python-socketio==5.10.0
httpx==0.25.2
numpy==1.24.3
tiktoken==0.5.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1
redis==5.0.1
gitpython==3.1.40
docker==6.1.3
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
